Add explicit stopped flag and simplify remove
Problem: If a Ticker calls Stop() on itself while we are in the middle
of Advance() and firing that Ticker, we accidentally add it back to the
queue when Advance() resumes, even though it is Stopped.

Solution: Using the mock field as stopped and doing binary search on the
(typically small) list of timers is too clever.  For now, just use a
stopped flag.