Fix unit test failure by not assuming timer handlers run to completion
Problem: The unit tests would sometimes fail because they did not see
enough results in the "fired" array:

    --- FAIL: TestAdvance (0.00s)
        mock_test.go:90: In stage 3, expected 4 fired found 3 ([3 2 4])
    FAIL
    FAIL    github.com/aspenmesh/tock       0.024s

Solution: These tests were assuming that the timer handler (select loop)
would notice and update the "fired" array before we checked it.  That is
invalid - it may be an indeterminate amount of time.  Rewrite the test
so there is a "sync group" to guarantee that the handler has completed
before we examine the results.  We will still verify the order of the
results.