Add Yield as an optional argument, the best tests shouldn't need
Problem: The logic to yield when firing a timer may feel more familiar
to users (timers often run at the time they are due, and to completion)
but it can hide dependence on timer handlers completing by a certain
time (which is not guaranteed).

Solution: Make this behavior an option that defaults to false.